Inventory & Shop System
This is the code of a prototype of an inventory system and a shop system that I wrote for the game "Arthain." Both systems inherit from the same script, "ItemsPanel," and have specific functions depending on which you are using. You select items through buttons that call from the "ItemSlot" script, and each button stores a "ScriptObj_InvItem" that contains the information on the item stored. For the shop system, it also uses the "GameManager" to increase/decrease the currency, depending on whether you are buying or selling.